move project.h/c from their own folder, no need right now
